refactor(database): split modular API families
