+++ packages/conformance/registry/rtdb.ts
@@ -178,9 +178,9 @@ const row19 = defineRows({
     api: "Modular SDK surface",
     status: "diverged-documented",
     statusNote: "logical disconnect lifecycle is modeled, but the in-memory data plane itself remains available",
-    evidence: "`unit:modular/on-disconnect.test.ts` + `unit:modular/fruit-aliases.test.ts`",
+    evidence: "`unit:database/on-disconnect.test.ts` + `unit:modular/fruit-aliases.test.ts`",
     automation: "unit-backed",
-    conformanceTests: ["packages/pyric/test/database/modular/on-disconnect.test.ts","packages/pyric/test/database/modular/fruit-aliases.test.ts"],
+    conformanceTests: ["packages/pyric/test/database/on-disconnect.test.ts","packages/pyric/test/database/modular/fruit-aliases.test.ts"],
   },
 });
 
@@ -543,7 +543,7 @@ export const rtdbRegistry = {
     },
     {
       kind: 'table',
-      prefix: "## Modular SDK surface\n\n`pyric/database` is the sandbox-only mirror of `firebase/database`'s\ntree-shakable free-function shape (`getDatabase`, `ref`, `child`, `get`,\n`set`, `update`, `remove`, `push`, listeners, queries, transactions, and\nsentinels). Package resolution selects production or sandbox before either\nmodule loads; this mirror has no production target or runtime dependency on\n`firebase/database`.\n\nTwo sandbox identity modes are selected by the value passed to\n`getDatabase`:\n\n- **Sandbox** — `getDatabase(ctx: SandboxContext)`. Frozen identity baked into\n  the handle at construction.\n- **Sandbox-live** — `getDatabase(sandbox: Sandbox)`. Identity read per\n  operation from `sandbox.currentUser`.\n\nThe implementation lives in `packages/pyric/src/database/modular.ts`; the\nin-process backend lives in `packages/pyric/src/database/sandbox/`. Rows below\nare scoped to the modular mirror. The pure rules-engine rows above are internal\ntooling coverage and are not exports of `pyric/database`.\n",
+      prefix: "## Modular SDK surface\n\n`pyric/database` is the sandbox-only mirror of `firebase/database`'s\ntree-shakable free-function shape (`getDatabase`, `ref`, `child`, `get`,\n`set`, `update`, `remove`, `push`, listeners, queries, transactions, and\nsentinels). Package resolution selects production or sandbox before either\nmodule loads; this mirror has no production target or runtime dependency on\n`firebase/database`.\n\nTwo sandbox identity modes are selected by the value passed to\n`getDatabase`:\n\n- **Sandbox** — `getDatabase(ctx: SandboxContext)`. Frozen identity baked into\n  the handle at construction.\n- **Sandbox-live** — `getDatabase(sandbox: Sandbox)`. Identity read per\n  operation from `sandbox.currentUser`.\n\nThe public implementation is split by API family under `packages/pyric/src/database/`; the\nin-process backend lives in `packages/pyric/src/database/sandbox/`. Rows below\nare scoped to the modular mirror. The pure rules-engine rows above are internal\ntooling coverage and are not exports of `pyric/database`.\n",
       rows: [
         row5({
           rowRef: "M1",
@@ -1635,24 +1635,24 @@ export const rtdbRegistry = {
           featureKeys: ["onDisconnect"],
           behavior: "`onDisconnect(ref)` returns the Firebase-shaped five-method handle; registration methods and `cancel()` return acknowledged Promises",
           status: "conforms",
-          evidence: "Oracle `rtdb-modular-ondisconnect-registration`; sandbox `unit:modular/on-disconnect.test.ts`.",
+          evidence: "Oracle `rtdb-modular-ondisconnect-registration`; sandbox `unit:database/on-disconnect.test.ts`.",
           risk: ["structural"],
           riskReasons: ["structural / routing-only claim"],
           automation: "oracle-backed",
           oracleObservations: ["rtdb-modular-ondisconnect-registration"],
-          conformanceTests: ["packages/pyric/test/database/modular/on-disconnect.test.ts"],
+          conformanceTests: ["packages/pyric/test/database/on-disconnect.test.ts"],
           rowNumber: 77,
         }),
         row5({
           rowRef: "M78",
           surface: "rtdb-modular",
-          featureKeys: ["onDisconnect","set","update","remove"],
-          behavior: "Acknowledged disconnect registration does not mutate server data before the client disconnects",
+          featureKeys: ["onDisconnect","set"],
+          behavior: "An acknowledged `onDisconnect(ref).set(value)` registration does not mutate server data before the client disconnects",
           status: "conforms",
           evidence: "Oracle `rtdb-modular-ondisconnect-registration`; sandbox registration/no-mutation test.",
           automation: "oracle-backed",
           oracleObservations: ["rtdb-modular-ondisconnect-registration"],
-          conformanceTests: ["packages/pyric/test/database/modular/on-disconnect.test.ts"],
+          conformanceTests: ["packages/pyric/test/database/on-disconnect.test.ts"],
           rowNumber: 78,
         }),
         row5({
@@ -1667,7 +1667,7 @@ export const rtdbRegistry = {
           riskReasons: ["asserts listener semantics"],
           automation: "oracle-backed",
           oracleObservations: ["rtdb-modular-ondisconnect-clean-set"],
-          conformanceTests: ["packages/pyric/test/database/modular/on-disconnect.test.ts"],
+          conformanceTests: ["packages/pyric/test/database/on-disconnect.test.ts"],
           rowNumber: 79,
         }),
         row5({
@@ -1679,7 +1679,7 @@ export const rtdbRegistry = {
           evidence: "Oracle `rtdb-modular-ondisconnect-operations-cancel`; sandbox operation/cancellation test, including the captured parent-set + child-cancel merge result.",
           automation: "oracle-backed",
           oracleObservations: ["rtdb-modular-ondisconnect-operations-cancel"],
-          conformanceTests: ["packages/pyric/test/database/modular/on-disconnect.test.ts"],
+          conformanceTests: ["packages/pyric/test/database/on-disconnect.test.ts"],
           rowNumber: 80,
         }),
         row5({
@@ -1694,7 +1694,7 @@ export const rtdbRegistry = {
           riskReasons: ["asserts rules-denial behavior"],
           automation: "oracle-backed",
           oracleObservations: ["rtdb-modular-ondisconnect-rules"],
-          conformanceTests: ["packages/pyric/test/database/modular/on-disconnect.test.ts"],
+          conformanceTests: ["packages/pyric/test/database/on-disconnect.test.ts"],
           rowNumber: 81,
         }),
         row5({
@@ -1703,9 +1703,9 @@ export const rtdbRegistry = {
           featureKeys: ["onDisconnect","goOffline"],
           behavior: "**Local lifecycle contract:** disconnect queues are owned per Database client, drain on app deletion, clear without executing on sandbox reset, and are excluded from persisted RTDB snapshots",
           status: "unverified",
-          evidence: "`unit:modular/on-disconnect.test.ts` covers independent clients, app deletion, reset clearing, and snapshot exclusion; these sandbox-owned boundaries have no production oracle assertion in this evidence set.",
+          evidence: "`unit:database/on-disconnect.test.ts` covers independent clients, app deletion, reset clearing, and snapshot exclusion; these sandbox-owned boundaries have no production oracle assertion in this evidence set.",
           automation: "unit-backed",
-          conformanceTests: ["packages/pyric/test/database/modular/on-disconnect.test.ts"],
+          conformanceTests: ["packages/pyric/test/database/on-disconnect.test.ts"],
           rowNumber: 82,
         }),
         row5({
@@ -1714,13 +1714,13 @@ export const rtdbRegistry = {
           featureKeys: ["onDisconnect"],
           behavior: "**Divergence:** `OnDisconnect.setWithPriority(value, priority)` writes `value` on disconnect but the sandbox does not retain RTDB priority metadata, so `snapshot.priority` remains `null`",
           status: "diverged-documented",
-          evidence: "Oracle `rtdb-modular-ondisconnect-operations-cancel` captures export-format priority 7; sandbox `unit:modular/on-disconnect.test.ts` pins the value-write plus null-priority boundary.",
+          evidence: "Oracle `rtdb-modular-ondisconnect-operations-cancel` captures export-format priority 7; sandbox `unit:database/on-disconnect.test.ts` pins the value-write plus null-priority boundary.",
           risk: ["specific-field"],
           riskScore: 1,
           riskReasons: ["asserts a specific field/property value"],
           automation: "oracle-backed",
           oracleObservations: ["rtdb-modular-ondisconnect-operations-cancel"],
-          conformanceTests: ["packages/pyric/test/database/modular/on-disconnect.test.ts"],
+          conformanceTests: ["packages/pyric/test/database/on-disconnect.test.ts"],
           rowNumber: 83,
         }),
         row5({
@@ -1729,7 +1729,7 @@ export const rtdbRegistry = {
           featureKeys: ["onDisconnect"],
           behavior: "**Divergence (pending fix):** clean `goOffline`, app deletion, playground `pagehide`, and best-effort MessagePort close drain queued operations, but unannounced total renderer/process loss is not guaranteed by the in-memory sandbox",
           status: "diverged-documented",
-          evidence: "Oracle `rtdb-modular-ondisconnect-abrupt-exit` proves Firebase executes an acknowledged registration after forced writer termination. The two-port worker integration proves deterministic cleanup; browser MessagePort close delivery cannot prove total-process loss without durable host-owned leases.",
+          evidence: "Oracle `rtdb-modular-ondisconnect-abrupt-exit` proves Firebase executes an acknowledged registration after forced writer termination. The two-port worker integration exercises `goOffline`, served-app deletion, and non-persisted `pagehide`; browser MessagePort close delivery cannot prove total-process loss without durable host-owned leases.",
           risk: ["listener"],
           riskScore: 2,
           riskReasons: ["asserts listener semantics"],
